Briefly started binary tree class and fixed Node class
My method of deleting and adding nodes before would not have worked, I have implemented a correct solution. I also added a return type of Node* to Del(). If the tree's parent gets replaced, it returns the pointer to the new parent so that the binary tree class can update it, otherwise returns NULL